+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `vat24api-php` will be documented in this file.
 
+## 1.0.0 - 2025-05-21
+
+- Initial release
+- Support for VAT number validation
+- Support for EORI number validation